% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchbinary.R
\name{covmatch.binary}
\alias{covmatch.binary}
\title{Covariate Matching (Match - Two)}
\usage{
covmatch.binary(dname, cov, weight, cov_circ)
}
\arguments{
\item{dname}{This should always be a list, containing all the data sets to match.}

\item{cov}{Vector of column numbers for the covariates to be used in matching.}

\item{weight}{Vector of threshold values, against which matching happens.
It can be a single value such as 0.2, or vector such as c(0.2, 0.2, 0.3),
considering three covariates.}

\item{cov_circ}{Vector stating the column position of circular variables such as wind direction,
nacelle position etc.}
}
\value{
The function returns a list containing after matched data sets.
}
\description{
The function aims to take exactly two data sets and returns the after
matched data sets using user specified covariates.
}
\examples{

fname <- list(data1, data2)
cov <- c(6, 1, 13)
weight <- c(0.1, 0.1, 0.5)

matched_data = covmatch.binary(fname, cov = cov, weight = weight)
}
