% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmany.R
\name{covmatch.mult}
\alias{covmatch.mult}
\title{Covariate Matching (Match - Many)}
\usage{
covmatch.mult(dname, cov, weight, cov_circ)
}
\arguments{
\item{dname}{This should always be a list, containing all the data sets to match.}

\item{cov}{Vector of column numbers for the covariates to be used in matching.}

\item{weight}{Vector of threshold values, against which matching happens.
It should be a vector such as c(0.2, 0.2, 0.3), if considering three covariates.}

\item{cov_circ}{Vector stating the column position of circular variables such as wind direction,
nacelle position etc.}
}
\value{
The function returns a list containing after matched data sets.
}
\description{
The function aims to take different data sets and returns the after
matched data sets using user specified covariates.
}
\examples{

fname <- list(Season1, Season2, Season3)
cov <- c(1, 6, 14)
weight <- c(0.2, 0.2, 0.2)

matched_data = covmatch.mult(fname, cov = cov, weight = weight)
}
